backend/pyproject.toml
[tool.poetry]
name = "data-extracter"
version = "0.1.0"
description = ""
authors = ["Caio <caiaoviniciuscc@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
python-multipart = "^0.0.20"
torch = "^2.0.0"
easyocr = "^1.7.2"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pymupdf = "^1.25.3"
opencv-python = "^4.11.0.86"
ocrmypdf = "^16.9.0"
pandas = "^2.2.3"
python-docx = "^1.1.2"
openai = "^1.65.1"
transformers = "^4.49.0"
cmake = "^3.31.6"
sentencepiece = "^0.2.0"
pyspellchecker = "^0.8.2"
keras-ocr = "^0.9.3"
dotenv = "^0.9.9"


[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
